# Optimize Http::Cookie::ToString()