Оптимизация отправки cookies с помощью SmallString::resize_and_overwrite().
```
-----------------------------------------------------------------------------------------
Benchmark                                               Time             CPU   Iterations
-----------------------------------------------------------------------------------------
http_cookie_serialization_append                      544 ns          544 ns      1244345
http_cookie_serialization_resize_and_overwrite        462 ns          462 ns      1489183
```